+++ src/events-store.ts
@@ -1,18 +1,23 @@
 export class EventsStore {
-  private static events: Record<string, () => void> = {};
+  private static events: Record<string, (arg?: any) => void> | undefined;
 
-  static set(name: string, fn: () => void) {
+  static set(name: string, fn: (arg?: any) => void) {
     if (!name) return;
-    this.events[name] = fn;
+    const eventsObj = { ...this.events, [name]: fn };
+    this.events = eventsObj;
   }
 
   static get(name: string) {
-    if (!name) return;
+    if (!name || !this.events) return;
     return this.events[name];
   }
 
-  static run(name: string) {
-    if (!name) return;
-    this.events[name]?.();
+  static run(name: string, arg?: any) {
+    if (!name || !this.events) return;
+    this.events[name]?.(arg);
+  }
+
+  static clear() {
+    this.events = undefined;
   }
 }