+++ src/types.ts
@@ -1,5 +1,7 @@
 import { ExtensionProvider } from '@elrondnetwork/erdjs-extension-provider';
+import { Transaction } from '@elrondnetwork/erdjs/out/transaction';
 import { WalletConnectProvider } from '@elrondnetwork/erdjs-wallet-connect-provider';
+import { WalletProvider } from '@elrondnetwork/erdjs-web-wallet-provider/out';
 
 export interface InitOptions {
   apiUrl: string;
@@ -8,6 +10,8 @@ export interface InitOptions {
   onLoginPending?: () => void;
   onLoggedIn?: () => void;
   onLogout?: () => void;
+  onTxStarted?: (transaction: Transaction) => void;
+  onTxFinalized?: (transaction: Transaction) => void;
 }
 
 export enum LoginMethodsEnum {
@@ -20,9 +24,11 @@ export enum LoginMethodsEnum {
 export type DappProvider =
   | ExtensionProvider
   | WalletConnectProvider
+  | WalletProvider
   | undefined;
 
 export interface LoginOptions {
   qrCodeContainer?: string | HTMLElement;
   token?: string;
+  callbackRoute?: string;
 }